+++ docs/changelog.rst
@@ -3,6 +3,7 @@ Changelog
 =========
 
 * :feature:`-` Added StockPlay module
+* :feature:`-` Added `@protected` decorator
 
 * :release:`4.1.0 <2019-02-10>`
 * :support:`-` First documented release in awhile. Many new modules and tests have been added. See the git log if you so desire.